# perf: optimize functional options