Previously, we wrapped functional options with interface (`interface { apply(o *options) }`) following the [Uber style](https://github.com/uber-go/guide/blob/master/style.md#functional-options). While this style (arguably) provides more flexibility, this comes with a trade-off too: "arguments passed to interface's method typically be heap-allocated".

I don't think we need such flexibility in our code and [ordinary functional options](https://dave.cheney.net/2014/10/17/functional-options-for-friendly-apis) would be sufficient for our cases. Additionally, we make Decoder to reuse listeners' slices to reduce allocation needed. By doing this, we can get zero allocation when resetting Decoder and Encoder.

Here is the update and the benchmark regarding this change:

#### Decoder:
```js
goos: darwin
goarch: amd64
pkg: github.com/muktihari/fit/decoder
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
                          │   old.txt    │               new.txt                │
                          │    sec/op    │    sec/op     vs base                │
Reset/benchmark_New()-4      4.975µ ± 8%   4.869µ ± 22%        ~ (p=0.447 n=10)
Reset/benchmark_Reset()-4   177.60n ± 1%   49.34n ±  6%  -72.22% (p=0.000 n=10)
geomean                      939.9n        490.2n        -47.85%

                          │   old.txt    │                  new.txt                  │
                          │     B/op     │     B/op      vs base                     │
Reset/benchmark_New()-4     23.54Ki ± 0%   23.49Ki ± 0%    -0.20% (p=0.000 n=10)
Reset/benchmark_Reset()-4     64.00 ± 0%      0.00 ± 0%  -100.00% (p=0.000 n=10)
geomean                     1.213Ki                      ?                       ¹ ²
¹ summaries must be >0 to compute geomean
² ratios must be >0 to compute geomean

                          │  old.txt   │                 new.txt                 │
                          │ allocs/op  │ allocs/op   vs base                     │
Reset/benchmark_New()-4     8.000 ± 0%   6.000 ± 0%   -25.00% (p=0.000 n=10)
Reset/benchmark_Reset()-4   3.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
geomean                     4.899                    ?                       ¹ ²
¹ summaries must be >0 to compute geomean
² ratios must be >0 to compute geomean
```

#### Encoder:
```js
goos: darwin
goarch: amd64
pkg: github.com/muktihari/fit/encoder
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
                          │   old.txt   │               new.txt                │
                          │   sec/op    │    sec/op     vs base                │
Reset/benchmark_New()-4     24.01µ ± 2%   23.96µ ± 49%        ~ (p=0.579 n=10)
Reset/benchmark_Reset()-4   72.90n ± 1%   30.97n ±  3%  -57.51% (p=0.000 n=10)
geomean                     1.323µ        861.4n        -34.89%

                          │   old.txt    │                  new.txt                  │
                          │     B/op     │     B/op      vs base                     │
Reset/benchmark_New()-4     128.1Ki ± 0%   128.1Ki ± 0%    -0.02% (p=0.000 n=10)
Reset/benchmark_Reset()-4     24.00 ± 0%      0.00 ± 0%  -100.00% (p=0.000 n=10)
geomean                     1.733Ki                      ?                       ¹ ²
¹ summaries must be >0 to compute geomean
² ratios must be >0 to compute geomean

                          │  old.txt   │                 new.txt                 │
                          │ allocs/op  │ allocs/op   vs base                     │
Reset/benchmark_New()-4     7.000 ± 0%   6.000 ± 0%   -14.29% (p=0.000 n=10)
Reset/benchmark_Reset()-4   1.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
geomean                     2.646                    ?                       ¹ ²
¹ summaries must be >0 to compute geomean
² ratios must be >0 to compute geomean
```